requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.21.3
azure-storage-blob==12.24.1
openai==1.63.2
requests==2.31.0
markdown==3.7